requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
dj-database-url>=2.2,<3.0
